[Refactor] 내부에서 사용하는 schema와 FE와의 통신에 사용되는 schema 분리
