+++ frontend/streamlit_prototype/app.py
@@ -11,7 +11,7 @@ from src.database.postgresql import health_check
 from src.service.route.route_service import get_route
 from src.service.route.child_walk_route import get_child_friendly_route
 from src.repository.graph_repository_v2 import load_graph_v2 as load_graph
-from src.service.map_service import fetch_local_db_lines_optimized, fetch_local_db_points
+from frontend.streamlit_prototype.services.map_service import fetch_local_db_lines_optimized, fetch_local_db_points
 from src.service.banner_service import get_banner
 from streamlit.components.v1 import html as st_html
 from chatbot_app import init_session, chat_and_assign_weights, run_async
