+++ frontend/streamlit_prototype/components/map_view.py
@@ -1,8 +1,8 @@
 import pydeck as pdk
 import streamlit as st
-from src.service.map_service import (
+from frontend.streamlit_prototype.services.map_service import (
     fetch_local_db_points,
-    fetch_local_db_lines_optimized
+    fetch_local_db_lines_optimized,
 )
 
 LOCATIONS = {
