--- src/agent/tools/route_tool.py
@@ -1,14 +1,0 @@
-from typing import Any
-
-def recommend_route_with_engine(route_request: dict[str, Any]) -> dict[str, Any]:
-    """
-    route_engine을 호출하는 agent tool 예정 함수.
-
-    예정 흐름:
-    1. RouteRequest 검증
-    2. route_engine/route_orchestrator 호출
-    3. RouteResult 반환
-
-    현재 단계에서는 실제 route_engine 호출 구현 금지.
-    """
-    pass
