+++ src/interfaces/api/auth_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, Cookie, Response
 
 from src.interfaces.dependencies import get_auth_service
 from src.service.auth_service import AuthService
-from src.schema.auth_schema import AuthResponse
+from src.interfaces.schema.auth_schema import AuthResponse
 
 router = APIRouter(
     prefix="/api/auth",
