+++ src/interfaces/api/user_router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends
 from src.interfaces.dependencies import get_user_service
-from src.schema.user_schema import UuidResponse
+from src.interfaces.schema.user_schema import UuidResponse
 from src.service.user_service import UserService
 
 router = APIRouter(
