+++ src/repository/chat/chat_state_repository.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from src.database.valkey import get_valkey_db
 from src.schema.prewalk_schema import State
-from src.service.prewalk.chatbot_utils import PydanticUtils
+from src.agent.utils.chatbot_utils import PydanticUtils
 
 class ChatStateRepository:
     @staticmethod
