+++ src/repository/route/graph_repository.py
@@ -1,7 +1,7 @@
 import networkx as nx
 from sqlalchemy import text
 from src.database.postgresql import get_postgresql_db
-from src.service.route.path_utils import remove_dead_ends
+from src.route_engine.path_utils import remove_dead_ends
 
 
 def load_graph() -> nx.Graph:
