+++ src/service/login_service.py
@@ -7,8 +7,8 @@ from urllib.parse import urlencode
 from src.service.user_service import UserService
 from src.service.auth_service import AuthService
 from src.repository.user.user_repository import UserRepository
-from src.schema.login_schema import LoginUrlResponse
-from src.schema.auth_schema import Status
+from src.interfaces.schema.login_schema import LoginUrlResponse
+from src.interfaces.schema.auth_schema import Status
 
 load_dotenv()
